Create project
I apologize for the lack of comments. I used setup() as a test bench for the methods of the Complex and Polar classes. Each Complex object has a real part and imaginary part (a+bi). We can take the real and imaginary parts and use trigonometry to find the angle and magnitude of the line they would represent on a coordinate plane.